IA Phase 3: unified People (Friends + Groups) inside Community
The 👥 People button now opens an inline People screen with Friends/Groups
sub-tabs, instead of routing away to /friends and /groups.

- Extracted FriendsPanel + GroupsPanel (the profile-panel pattern); /friends and
  /groups routes are now thin wrappers around them.
- Community ▸ People swaps the main tab bar for a "← Community" back + Friends ·
  Groups sub-tabs, rendering the panels inline (full group detail: wealth/compete,
  members, add-friends, chat — all in-hub).
- Friends list gains ⚔️ challenge (opens the New-Challenge builder pre-filled with
  that friend) + a tappable name → profile.

Verified: QA 19/19, 0 console errors; live People flow (search, create group,
group detail, back nav) all work in-hub. Routes still work standalone.
